sagemaker/requirements.txt
boto3==1.26.44
sagemaker==2.127.0
snowflake-connector-python==2.9.0